+++ api/config/config_prod.json
@@ -107,7 +107,6 @@
     "exportObjectsPerTask" : 10
   },
   "captcha": {
-    "enableCaptcha": false,
-    "serverKeyFileName": "captcha-server-key.txt"
+    "enableCaptcha": false
   }
 }
