+++ api/config/config_staging.json
@@ -107,7 +107,6 @@
     "exportObjectsPerTask" : 10
   },
   "captcha": {
-    "enableCaptcha": false,
-    "serverKeyFileName": "captcha-server-key.txt"
+    "enableCaptcha": false
   }
 }
