+++ common-api/src/main/java/org/pmiops/workbench/config/WorkbenchConfig.java
@@ -256,6 +256,5 @@ public class WorkbenchConfig {
 
   public static class CaptchaConfig {
     public boolean enableCaptcha;
-    public String serverKeyFileName;
   }
 }
